@@ -6,8 +6,8 @@ import {
   Grid,
   TextField,
   Typography,
-} from "@material-ui/core";
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
+} from "@mui/material";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 
 import { Savefile } from "../../interfaces";
 
